WIP: NOT WORKING - continued refactoring
